refactor: Product 관련 UI 파일 위치 이동
