+++ app/src/main/java/woowacourse/shopping/ui/product/ProductAdapter.kt
@@ -1,0 +1,24 @@
+package woowacourse.shopping.ui.product
+
+import android.view.ViewGroup
+import androidx.recyclerview.widget.RecyclerView
+import woowacourse.shopping.model.Product
+
+class ProductAdapter(
+    private val items: List<Product>,
+    private val onClickProduct: (Product) -> Unit,
+) : RecyclerView.Adapter<ProductViewHolder>() {
+    override fun onCreateViewHolder(
+        parent: ViewGroup,
+        viewType: Int,
+    ): ProductViewHolder = ProductViewHolder.from(parent, onClickProduct)
+
+    override fun onBindViewHolder(
+        holder: ProductViewHolder,
+        position: Int,
+    ) {
+        holder.bind(items[position])
+    }
+
+    override fun getItemCount(): Int = items.size
+}
