+++ app/src/main/res/layout/activity_main.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="vm"
-            type="woowacourse.shopping.ui.MainViewModel" />
+            type="woowacourse.shopping.ui.product.MainViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
