+++ app/src/test/java/woowacourse/shopping/ui/product/MainActivityTest.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping
+package woowacourse.shopping.ui.product
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.lifecycle.ViewModelProvider
@@ -8,23 +8,20 @@ import org.junit.Test
 import org.junit.runner.RunWith
 import org.robolectric.Robolectric
 import org.robolectric.RobolectricTestRunner
-import woowacourse.shopping.ui.MainActivity
-import woowacourse.shopping.ui.MainViewModel
-
 
 @RunWith(RobolectricTestRunner::class)
 class MainActivityTest {
-
     @get:Rule
     var instantTaskExecutorRule = InstantTaskExecutorRule()
 
     @Test
     fun `Activity 실행 테스트`() {
         // given
-        val activity = Robolectric
-            .buildActivity(MainActivity::class.java)
-            .create()
-            .get()
+        val activity =
+            Robolectric
+                .buildActivity(MainActivity::class.java)
+                .create()
+                .get()
 
         // then
         assertThat(activity).isNotNull()
@@ -33,10 +30,11 @@ class MainActivityTest {
     @Test
     fun `ViewModel 주입 테스트`() {
         // given
-        val activity = Robolectric
-            .buildActivity(MainActivity::class.java)
-            .create()
-            .get()
+        val activity =
+            Robolectric
+                .buildActivity(MainActivity::class.java)
+                .create()
+                .get()
         val viewModel = ViewModelProvider(activity)[MainViewModel::class.java]
 
         // then
